Refactor RÚNATAL frame scanning
